+++ database/migrations/2022_06_22_100334_create_gostaresh_changes_number_of_research_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->foreignId('city_id')->nullable();
             $table->foreignId('rural_district_id')->nullable();
 
-            $table->unsignedBigInteger('number_of_research')->default(0);
+            $table->float('number_of_research')->default(0);
 
             $table->integer('year')->unsigned()->nullable();
             $table->tinyInteger('month')->unsigned()->nullable();
