+++ database/migrations/2022_06_23_072229_create_gostaresh_average_tuition_incomes_table.php
@@ -31,11 +31,11 @@ return new class extends Migration
             $table->string('unit')->nullable();
             $table->string('university_type')->comment('دانشگاه')->nullable();
             $table->unsignedTinyInteger('department_of_education')->nullable();
-            $table->unsignedInteger('associate_degree')->nullable();
-            $table->unsignedInteger('bachelor_degree')->nullable();
-            $table->unsignedInteger('masters')->nullable();
-            $table->unsignedInteger('professional_phd')->nullable();
-            $table->unsignedInteger('phd')->nullable();
+            $table->float('associate_degree')->nullable();
+            $table->float('bachelor_degree')->nullable();
+            $table->float('masters')->nullable();
+            $table->float('professional_phd')->nullable();
+            $table->float('phd')->nullable();
 
             $table->integer('year')->unsigned()->nullable();
             $table->tinyInteger('month')->unsigned()->nullable();
